+++ CHANGELOG.md
@@ -430,3 +430,4 @@
 * Remove `|>` existential polymorphic type and its Substructure instance
 * Remove `lookup` and `discover` methods from `Dictionary` interface
 * Define `lookup` method in `Morphable` ability module
+* Remove `binary` expression from `Binary` module